now neutral subtraction works
